+++ extensions/gii/generators/crud/form.php
@@ -6,8 +6,8 @@
 echo $form->field($generator, 'modelClass');
 echo $form->field($generator, 'searchModelClass');
 echo $form->field($generator, 'controllerClass');
+echo $form->field($generator, 'viewPath');
 echo $form->field($generator, 'baseControllerClass');
-echo $form->field($generator, 'moduleID');
 echo $form->field($generator, 'indexWidgetType')->dropDownList([
     'grid' => 'GridView',
     'list' => 'ListView',