+++ tests/unit/framework/widgets/MenuTest.php
@@ -3,7 +3,6 @@
 namespace yiiunit\framework\widgets;
 
 use yii\widgets\Menu;
-use yii\widgets\Spaceless;
 
 /**
  * @group widgets